refactor public route rendering and auth gates
