+++ src/lib/auth.tsx
@@ -1,4 +1,5 @@
-import { createContext, useContext, useEffect, useState } from 'react';
+import type { ReactNode } from "react";
+import { createContext, useContext } from "react";
 import { useConvexAuth } from "convex/react";
 import { useUser } from "@clerk/nextjs";
 
@@ -21,27 +22,14 @@ interface AuthProviderProps {
 export function AuthProvider({ children }: AuthProviderProps) {
   const { isAuthenticated, isLoading } = useConvexAuth();
   const { user: clerkUser, isLoaded: clerkLoaded } = useUser();
-  const [userId, setUserId] = useState<string | null>(null);
-  
-  // Obter o ID real do usuário do Clerk
-  useEffect(() => {
-    if (isAuthenticated && clerkUser && clerkLoaded && !userId) {
-      // Em produção, usamos o ID real do usuário do Clerk
-      setUserId(clerkUser.id);
-      
-      // Log para depuração (pode ser removido em produção)
-
-    } else if (!isAuthenticated) {
-      setUserId(null);
-    }
-  }, [isAuthenticated, clerkUser, clerkLoaded, userId]);
+  const userId = isAuthenticated && clerkLoaded && clerkUser ? clerkUser.id : null;
 
   return (
-    <AuthContext.Provider value={{ 
-      isAuthenticated, 
-      userId, 
+    <AuthContext.Provider value={{
+      isAuthenticated,
+      userId,
       // Consideramos que estamos carregando se o Clerk ainda estiver carregando OU se o Convex ainda está verificando
-      isLoading: isLoading || !clerkLoaded 
+      isLoading: isLoading || !clerkLoaded,
     }}>
       {children}
     </AuthContext.Provider>
@@ -51,7 +39,7 @@ export function AuthProvider({ children }: AuthProviderProps) {
 export function useAuth() {
   const context = useContext(AuthContext);
   if (context === undefined) {
-    throw new Error('useAuth must be used within an AuthProvider');
+    throw new Error("useAuth must be used within an AuthProvider");
   }
   return context;
-} 
+}
