+++ src/lib/providers/QueryProvider.tsx
@@ -1,5 +1,6 @@
-'use client';
+"use client";
 
+import type { ReactNode } from "react";
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { useState } from 'react';
