+++ src/lib/services/activityService.ts
@@ -6,7 +6,8 @@ import type { Id } from "@/../convex/_generated/dataModel";
 import { useCurrentUser } from "@/lib/hooks/useCurrentUser";
 
 // Helper function to query tickets - this will get tickets for an activity
-function queryTickets(): Promise<ActivityTicket[]> {
+function queryTickets(activityId: string): Promise<ActivityTicket[]> {
+  void activityId;
   return new Promise((resolve) => {
     // This is just a placeholder implementation
     // In a real app, you'd fetch the tickets from your backend
@@ -502,10 +503,10 @@ export const useCreateActivityTicket = () => {
       ...ticketData,
       availableQuantity: ticketData.availableQuantity,
       maxPerOrder: ticketData.maxPerOrder,
-    };
+    } as Parameters<typeof createTicketMutation>[0];
     
     try {
-      const ticketId = await createTicketMutation(convexData as any);
+      const ticketId = await createTicketMutation(convexData);
       return ticketId;
     } catch (error) {
       console.error("Error creating ticket:", error);
@@ -520,11 +521,13 @@ export const useUpdateActivityTicket = () => {
   return async (ticketData: ActivityTicket) => {
     const { id, createdAt: _createdAt, ...updateData } = ticketData;
     void _createdAt; // Avoid unused variable warning
-    
-    return await updateTicketMutation({
+
+    const mutationInput = {
       id: id as Id<"activityTickets">,
       ...updateData,
-    } as any);
+    } as Parameters<typeof updateTicketMutation>[0];
+    
+    return await updateTicketMutation(mutationInput);
   };
 };
